+++ main.go
@@ -8,8 +8,8 @@ import (
 	"node-agent/pkg/applicationprofilemanager"
 	applicationprofilemanagerv1 "node-agent/pkg/applicationprofilemanager/v1"
 	"node-agent/pkg/config"
-	"node-agent/pkg/containerwatcher/dnsmanager"
 	"node-agent/pkg/containerwatcher/v1"
+	"node-agent/pkg/dnsmanager"
 	"node-agent/pkg/filehandler/v1"
 	"node-agent/pkg/networkmanager"
 	"node-agent/pkg/relevancymanager"
