+++ pkg/containerwatcher/v1/container_watcher.go
@@ -6,7 +6,7 @@ import (
 	"node-agent/pkg/applicationprofilemanager"
 	"node-agent/pkg/config"
 	"node-agent/pkg/containerwatcher"
-	"node-agent/pkg/containerwatcher/dnsmanager"
+	"node-agent/pkg/dnsmanager"
 	"node-agent/pkg/networkmanager"
 	"node-agent/pkg/relevancymanager"
 	"node-agent/pkg/utils"
