+++ pkg/dnsmanager/dns_manager_interface.go
